requirements.txt
fastapi
uvicorn
tensorflow-cpu
pillow
gunicorn
requests
python-multipart
flask
gradio
Flask
CORS
flask-cors